backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2
djangorestframework-simplejwt
openai==0.28
numpy